+++ lib/stellar_flutter_sdk.dart
@@ -178,6 +178,10 @@ export 'src/sep/0008/regulated_assets.dart';
 /// Soroban
 export 'src/soroban/soroban_client.dart';
 export 'src/soroban/soroban_server.dart';
+export 'src/soroban/soroban_rpc_requests.dart';
+export 'src/soroban/soroban_rpc_responses.dart';
+export 'src/soroban/soroban_transaction_responses.dart';
+export 'src/soroban/soroban_ledger_event_responses.dart';
 export 'src/soroban/soroban_auth.dart';
 export 'src/soroban/soroban_contract_parser.dart';
 export 'src/soroban/soroban_passkey.dart';
